Item manager update.
Access to the singleton was included in the item manager script. Furthermore, the coin counting problem was solved; it was enough to change the coin quantity variable in the `Add coins` function from an integer equal to 1 to a floating-point number equal to 0.5. Because if I can't understand why it's adding double, I'll force the logic to add half of that, to get 1, until I figure out the reason for the error. But it's solved.