Implementation of the weapon collection and projectile firing system.
Prefabs were created for the projectiles with a container, so that the animation made in the animator could work for both right and left-side shots. In addition, a Weapon folder was created to include 3 scripts for the weapon collection and firing system. The scripts are: InventoryBase (to allow weapon collection and equipping), Projectile Base (for firing mechanics), and weapon base (to create the weapon). Visual effects were included in the weapon's collectible icon, with a particle system.