Update of the player, health and destroy helper scripts
Minor adjustments to the code to try and prevent the player from shooting themselves. With the implementation of the scriptable objects for the player variants, unfortunately the player can no longer activate the weapon sprite and without it, they cannot shoot. When manually activated in the Unity inspector, it does not recognize the character's direction and always shoots in the same direction. This did not occur before; I have to fix this bug.